Add a workaround for tsk
Tsk[1] is a tool to record battles of Soku. It may record a battle times
when used with Giuroll, since rollbacks may happen when KO, or even the
result of the battle can changed because of rollback. This commit uses a
"fake" CBattleManager structure to provide stable fields for tsk.

[1] Tsk: https://wikiwiki.jp/thtools/%E5%A4%A9%E5%89%87%E8%A6%B3